xml/parsing/manakai/doctypes-2.dat

#data
<!DOCTYPE a PUBLIC><a/>
#errors
1;1;no XML decl;;;s
1;19;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a PUBLIC ><a/>
#errors
1;1;no XML decl;;;s
1;20;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b"><a/>
#errors
1;1;no XML decl;;;s
1;23;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" ><a/>
#errors
1;1;no XML decl;;;s
1;24;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b>c"><a/>
#errors
1;1;no XML decl;;;s
1;22;parser:literal not closed;;;m
1;23;text outside of root element;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c>d"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "b" "c>d">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b""c"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;23;no space before literal;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b"'c'><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;23;no space before literal;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC 'b'"c"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;23;no space before literal;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC 'b''c'><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;23;no space before literal;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a SYSTEM><a/>
#errors
1;1;no XML decl;;;s
1;19;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a SYSTEM ><a/>
#errors
1;1;no XML decl;;;s
1;20;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a SYSTEM "b"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a SYSTEM "b>c"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "" "b>c">
| <{} a>

#data
<!DOCTYPE a public "b" "c"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;13;keyword-wrong-case;;public;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIc "b" "c"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;13;keyword-wrong-case;;PUBLIc;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a system "b"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;13;keyword-wrong-case;;system;m
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a SYSTEm "b"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;13;keyword-wrong-case;;SYSTEm;m
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPEabc><abc/>
#errors
1;1;no XML decl;;;s
1;10;bogus DOCTYPE;;;m
#document
| <!DOCTYPE abc>
| <{} abc>

#data
<!DOCTYPE[]><a/>
#errors
1;1;no XML decl;;;s
1;10;bogus DOCTYPE;;;m
1;10;no DOCTYPE name;;;m
#document
| <!DOCTYPE >
| <{} a>

#data
<!DOCTYPE []><a/>
#errors
1;1;no XML decl;;;s
1;11;no DOCTYPE name;;;m
#document
| <!DOCTYPE >
| <{} a>

#data
<!DOCTYPE a[]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a []><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a PUBLIC[]><a/>
#errors
1;1;no XML decl;;;s
1;19;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a PUBLIC []><a/>
#errors
1;1;no XML decl;;;s
1;20;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a PUBLIC "[]"><a/>
#errors
1;1;no XML decl;;;s
1;24;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a "[]" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b"[]><a/>
#errors
1;1;no XML decl;;;s
1;23;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" []><a/>
#errors
1;1;no XML decl;;;s
1;24;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "[]"><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "b" "[]">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c"[]><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c" []><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c" d[]><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;28;bogus DOCTYPE;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c" de[]><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;28;bogus DOCTYPE;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" c[]><a/>
#errors
1;1;no XML decl;;;s
1;24;bogus DOCTYPE;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c" d[]e><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;28;bogus DOCTYPE;;;m
1;31;string after internal subset;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b" "c" d[]e[]f><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;28;bogus DOCTYPE;;;m
1;31;string after internal subset;;;m
#document
| <!DOCTYPE a "b" "c">
| <{} a>

#data
<!DOCTYPE a PUBLIC "b"[]"c"><a/>
#errors
1;1;no XML decl;;;s
1;23;no DOCTYPE literal;;;m
1;25;string after internal subset;;;m
#document
| <!DOCTYPE a "b" "">
| <{} a>

#data
<!DOCTYPE a SYSTEM[]><a/>
#errors
1;1;no XML decl;;;s
1;19;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a SYSTEM []><a/>
#errors
1;1;no XML decl;;;s
1;20;no DOCTYPE literal;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a SYSTEM "b"[]><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a SYSTEM "b" []><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a SYSTEM "b" c[]><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;24;bogus DOCTYPE;;;m
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a SYSTEM "b" cdefg[]><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;24;bogus DOCTYPE;;;m
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a SYSTEM "b" c[]d><a/>
#errors
1;1;no XML decl;;;s
1;1;external entref not expanded;;;w
1;24;bogus DOCTYPE;;;m
1;27;string after internal subset;;;m
#document
| <!DOCTYPE a "" "b">
| <{} a>

#data
<!DOCTYPE a [
#errors
1;1;no XML decl;;;s
1;14;parser:EOF;;;m
1;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[]
#errors
1;1;no XML decl;;;s
1;14;parser:EOF;;;m
1;14;parser:EOF;;;m
1;14;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[?]><a/>
#errors
1;1;no XML decl;;;s
1;13;string in internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a [][]><a/>
#errors
1;1;no XML decl;;;s
1;15;string after internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a [][>]><a/>
#errors
1;1;no XML decl;;;s
1;15;string after internal subset;;;m
1;17;text outside of root element;;;m
1;18;text outside of root element;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a []a[]><a/>
#errors
1;1;no XML decl;;;s
1;15;string after internal subset;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a []a[]
#errors
1;1;no XML decl;;;s
1;15;string after internal subset;;;m
1;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;3;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!UNKNOWN>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;3;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!E>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;4;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!EXT>
<!ATTLIST b>
]><a/>
#errors
1;1;no XML decl;;;s
2;4;bogus comment;;;m
3;1;empty attlist decl;;b;w
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ATTLIST b>
<!E
#errors
1;1;no XML decl;;;s
2;1;empty attlist decl;;b;w
3;4;bogus comment;;;m
3;4;parser:EOF;;;m
3;4;no root element;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
